submit_server/requirements.txt
Flask==1.0.2
Flask-OAuth==0.12
Flask-WTF==0.14.2
Flask-SQLAlchemy==2.3.2
cryptography==2.6.1
python-dotenv==0.10.1
Flask-APScheduler==1.11.0
Flask-Migrate==2.4.0
Flask-Table==0.5.0
plotly==3.7.1
